+++ libs/base/src/num/float/mod.rs
@@ -1,0 +1,19 @@
+// devela_base::num::float
+//
+//! Floating-point functionality.
+//
+
+mod alias; // fsize
+
+crate::items! { // structural access: _mods, _all
+    #[allow(unused)]
+    pub use _mods::*;
+
+    mod _mods { #![allow(unused)]
+        pub use super::alias::*;
+    }
+    pub(super) mod _all { #![allow(unused)]
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+}