+++ libs/base/src/num/int/alias.rs
@@ -1,4 +1,4 @@
-// devela::num::int::alias
+// devela_base::num::int::alias
 
 #![allow(non_camel_case_types)]
 
@@ -8,8 +8,8 @@ use crate::paste;
 macro_rules! iusize_alias {
     ($casted:literal, $sign:literal, $ty:ident, $($width:literal : $cast:ty),+) => { paste! {
         $(
-            #[doc = crate::TAG_PRIMITIVE!()]
-            #[doc = crate::TAG_NUM!()]
+            #[doc = $crate::TAG_PRIMITIVE!()]
+            #[doc = $crate::TAG_NUM!()]
             #[doc = $casted " pointer-sized " $sign " integer primitive."]
             #[cfg(target_pointer_width = $width)]
             pub type $ty = $cast;