+++ libs/base/src/num/int/mod.rs
@@ -1,0 +1,19 @@
+// devela_base::num::int
+//
+//! Integer functionality.
+//
+
+mod alias; // [i|u]size_[down|up]
+
+crate::items! { // structural access: _mods, _all
+    #[allow(unused)]
+    pub use _mods::*;
+
+    mod _mods { #![allow(unused)]
+        pub use super::alias::*;
+    }
+    pub(super) mod _all { #![allow(unused)]
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+}