+++ src/num/int/reexports.rs
@@ -1,0 +1,7 @@
+// devela::num::int::reexports
+//
+//! Reexported items from `devela_base`.
+//
+
+#[doc = crate::TAG_DEVELA_BASE!()]
+pub use devela_base::{isize_down, isize_up, usize_down, usize_up};